project1/requirements.txt
Flask==2.0.2
Flask-MySQL==1.5.1
Flask-JWT-Extended==4.1.0
stripe==2.56.0
python-dotenv==0.19.2
